add multi line select support
